refactor `index.js` && init separate function `addToCart`
